Add files via upload
This Addon allows the Odoo Instance to share the images across Image repos for different sites on the same Odoo instance / Server. This is particularly useful if you are hosting multiple sites on an Odoo 18 CE instance and want to share images without having to download / upload them multiple times. This addon is very handy if you're using the Unspalsh Odoo integration and more so if you use the Free Edition of Unsplash because it reduces the number of searches used. 

More saliently, this addon overcomes the Issue in Odoo 18CE where if an Image stored in an image repositiory for one site is used in another the image doesn't render (because permissions disallows the second site access to the image). 


This addon works well (And was developed for use in conjunction with) my Odoo Website Domain Router addon (Also called Odoo CE Multi Site), because that will enable an Odoo CE instance to hjost multiple websites rather than just the two which are enabled in Odoo 18 CE by default.